requirements.txt
emoji-country-flag==1.3.2
pandarallel==1.6.5
pandas==1.5.3
pycountry==19.8.18
requests==2.28.2
requests-oauthlib==1.3.1
beautifulsoup4==4.11.1
python-dotenv==1.0.0
